backend/schemas/user.py
from pydantic import BaseModel
from datetime import datetime
from uuid import UUID
from typing import Optional

class UserUpdate(BaseModel):
    first_name: Optional[str] = None
    last_name: Optional[str] = None
    phone: Optional[str] = None
    dob: Optional[datetime] = None
    role: Optional[str] = None
    profile_image: Optional[str] = None
    address: Optional[str] = None
    city: Optional[str] = None
    state: Optional[str] = None

class UserResponse(BaseModel):
    id: UUID
    email: str
    first_name: str
    last_name: str
    phone: Optional[str]
    role: str
    is_verified: bool
    created_at: datetime
    
    class Config:
        from_attributes = True



