backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
sqlalchemy==2.0.44
alembic==1.17.2
python-dotenv==1.2.1
asyncpg==0.31.0
pydantic-settings==2.12.0
psycopg2-binary==2.9.11
email-validator==2.3.0
PyJWT==2.10.1 
httpx==0.28.1